# Conflicts:
#	ui/package-lock.json
#	ui/package.json
#	ui/src/app/containers/AddForumPage/index.js
#	ui/src/app/containers/AddPostPage/index.js
#	ui/src/app/containers/AddThreadPage/index.js
#	ui/src/app/containers/ForumsPage/index.js

+++ NUStartApplication/NUStartApplication-ejb/src/java/session/GuideSessionBean.java
@@ -8,6 +8,7 @@ package session;
 import entity.Category;
 import entity.Comment;
 import entity.Guide;
+import entity.Link;
 import entity.Person;
 import error.NoResultException;
 import java.util.List;
@@ -43,8 +44,12 @@ public class GuideSessionBean implements GuideSessionBeanLocal {
     }
 
     @Override
-    public void createGuide(Guide g) {
+    public void createGuide(Long cId, Guide g) throws NoResultException {
+        g.setTitle(g.getTitle().trim());
+        g.setContent(g.getContent().trim());
+        Category c = getCategory(cId);
         em.persist(g);
+        c.getGuides().add(g);
     }
 
     @Override
@@ -53,17 +58,27 @@ public class GuideSessionBean implements GuideSessionBeanLocal {
         return q.getResultList();
     }
 
+    public Category getCategoryFromGuide(Long gId) throws NoResultException {
+        try {
+            Guide g = getGuide(gId);
+            Query q = em.createQuery("SELECT DISTINCT c FROM Category c WHERE :guide MEMBER OF c.guides");
+            q.setParameter("guide", g);
+
+            return (Category) q.getSingleResult();
+        } catch (Exception e) {
+            throw new NoResultException("Guide not found!");
+        }
+    }
+
     @Override
     public void updateGuide(Guide g) throws NoResultException {
         try {
             Guide oldG = getGuide(g.getId());
             oldG.setTitle(g.getTitle());
             oldG.setContent(g.getContent());
-//            if (g.getPublished()) {
-//                oldG.setDatePublished(new Date());
-//            } else {
-//                oldG.setDateUpdated(new Date());
-//            }
+            oldG.setPictureUrl(g.getPictureUrl());
+            oldG.setFiles(g.getFiles());
+            oldG.setLinks(g.getLinks());
         } catch (NoResultException ex) {
             Logger.getLogger(GuideSessionBean.class.getName()).log(Level.SEVERE, null, ex);
         }
@@ -71,9 +86,16 @@ public class GuideSessionBean implements GuideSessionBeanLocal {
     }
 
     @Override
-    public void deleteGuide(Long gId) throws NoResultException {
+    public void deleteGuide(Long cId, Long gId) throws NoResultException {
+        Category c = getCategory(cId);
         Guide g = getGuide(gId);
-        em.remove(g);
+
+        if (c != null) {
+            c.getGuides().remove(g);
+            em.remove(g);
+        } else {
+            throw new NoResultException("Not found");
+        }
     }
 
     @Override
@@ -132,9 +154,54 @@ public class GuideSessionBean implements GuideSessionBeanLocal {
     }
 
     @Override
-    public List<Comment> getCommentReplies(Long cId) {
+    public List<Comment> getCommentReplies(Long cId) throws NoResultException {
         Query q = em.createQuery("SELECT c FROM Comment c WHERE c.parent.id = :cId");
         q.setParameter("cId", cId);
         return q.getResultList();
+    }
+
+    @Override
+    public List<Comment> getComments(Long gId) throws NoResultException {
+        Guide guide = getGuide(gId);
+        return guide.getComments();
+    }
+
+    @Override
+    public Category getCategory(Long cId) throws NoResultException {
+        Category f = em.find(Category.class, cId);
+
+        if (f != null) {
+            return f;
+        } else {
+            throw new NoResultException("Not found");
+        }
+    }
+
+    @Override
+    public void addLink(Long gId, Link l) throws NoResultException {
+        Guide guide = getGuide(gId);
+        l.setName(l.getName().trim());
+        em.persist(l);
+        guide.getLinks().add(l);
+    }
+
+    @Override
+    public void editLink(Link l) throws NoResultException {
+        Link oldL = em.find(Link.class, l.getId());
+        oldL.setName(l.getName().trim());
+        oldL.setURL(l.getURL().trim());
+    }
+
+    @Override
+    public void deleteLink(Long gId, Long lId) throws NoResultException {
+        Guide g = getGuide(gId);
+        Link l = em.find(Link.class, lId);
+
+        if (g != null) {
+            g.getLinks().remove(l);
+            em.remove(l);
+        } else {
+            throw new NoResultException("Not found");
+        }
     }
 }
